% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_period.R
\name{get_period}
\alias{get_period}
\title{Get period information from an object}
\usage{
get_period(object)
}
\arguments{
\item{object}{a periodic object}
}
\description{
Get period information from an object
}
