% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggperiodic.R
\docType{package}
\name{ggperiodic}
\alias{ggperiodic}
\alias{ggperiodic-package}
\title{ggperiodic: Easy Plotting of Periodic Data with 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Implements methods to plot periodic data in any arbitrary range on the fly.
}
\section{Overview}{

The only thing you need to do is add the periodic information to a
data frame with \code{\link[=periodic]{periodic()}}. You then can manually wrap your data around any
domain with \code{\link[=wrap]{wrap()}} or just let ggplot2 do it automatically for you
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eliocamp/ggperiodic}
  \item Report bugs at \url{https://github.com/eliocamp/ggperiodic/issues}
}

}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})

}
