% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic.R
\name{periodic}
\alias{periodic}
\alias{periodic.default}
\alias{periodic.data.frame}
\alias{periodic.Layer}
\title{Add or remove periodic variables}
\usage{
periodic(object, ...)

\method{periodic}{default}(object, period, ...)

\method{periodic}{data.frame}(object, ...)

\method{periodic}{Layer}(object, ...)
}
\arguments{
\item{object}{the object to coherce to periodic}

\item{...}{name-value pairs of expressions defining the period}

\item{period}{a numeric vector whose range defines the period}
}
\description{
Creates a periodic object by specifying the periodic variables and their
periods.
}
\details{
This is a generic function and ggperiodic provides methods for vectors, data frames
and [ggplot2] Layers (geoms and stats).
}
