% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unperiodic.R
\name{unperiodic}
\alias{unperiodic}
\title{Remove periodic specifications}
\usage{
unperiodic(object, ...)
}
\arguments{
\item{object}{the object to remove periodicities}

\item{...}{arguments to methods}
}
\value{
An object of the same class as \code{object} but with no periodic subclass or
periodicity specifications.
}
\description{
Remove periodic specifications
}
