% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{wrap}
\alias{wrap}
\title{Wrap periodic data to an arbitary range}
\usage{
wrap(data, ..., .group = NULL)
}
\arguments{
\item{data}{a periodic data frame}

\item{...}{name-value pairs of expressions defining range specifications}

\item{.group}{optional group column}
}
\description{
Wrap periodic data to an arbitary range
}
